refactor(room): handle guest users references and show modal for new users only
