+++ src/components/Room/Modal/Footer.tsx
@@ -3,7 +3,7 @@ import { Button } from '@/components/ui/button';
 import { DialogFooter } from '@/components/ui/dialog';
 import LanguageDropdown from '@/components/LanguageDropdown';
 import { useState } from 'react';
-import { addDoc, collection, doc, setDoc, updateDoc } from 'firebase/firestore';
+import { collection, doc, setDoc, updateDoc } from 'firebase/firestore';
 import { useParticipantStore } from '@/store';
 import { db } from '@/firebase-config';
 import { useTranslation } from 'react-i18next';
@@ -35,10 +35,15 @@ const Footer = ({
     }
     setError('');
 
-    const roomRef = doc(db, 'rooms', roomUid);
-    const adminRef = doc(db, 'rooms', roomRef.id, 'participants', adminUserUid);
-
     if (isAdmin) {
+      const roomRef = doc(db, 'rooms', roomUid);
+      const adminRef = doc(
+        db,
+        'rooms',
+        roomRef.id,
+        'participants',
+        adminUserUid
+      );
       await updateDoc(roomRef, {
         uid: roomUid,
         name: roomName
@@ -52,15 +57,16 @@ const Footer = ({
       setParticipantId(adminUserUid);
     }
     if (!isAdmin) {
-      const participantRef = await addDoc(
-        collection(db, 'rooms', roomUid, 'participants'),
-        {
-          avatar: '',
-          name: newParticipantName,
-          vote: null
-        }
+      const participantDocRef = doc(
+        collection(db, 'rooms', roomUid, 'participants')
       );
-      setParticipantId(participantRef.id);
+      await setDoc(participantDocRef, {
+        uid: participantDocRef.id,
+        avatar: '',
+        name: newParticipantName,
+        vote: null
+      });
+      setParticipantId(participantDocRef.id);
     }
     setModalVisibility(false);
   };
